Split off suppliers to their own module.
DateFormatter now lives in a utility submodule of the 'suppliers' module.
