+++ suppliers/__init__.py
@@ -1,0 +1,4 @@
+from .hetgerecht import HetGerecht
+from .nomssupplier import NomsSupplier
+from .radboudumc import RadboudUMC
+from .refter import Refter
